rafactor objects overview
- dedupes various util functions
- splits everything into multiple components and hooks
- adds context menu for objects that allow for going to said object in the editor
