+++ electron-app/src/components/overview/objects/constants.ts
@@ -1,0 +1,22 @@
+export const LABEL_WIDTH = 208;
+export const ROW_HEIGHT = 44;
+export const HIDDEN_ROW_HEIGHT = 36;
+export const HIDDEN_ROW_VERTICAL_PADDING = 6;
+export const AXIS_HEIGHT = 30;
+export const MIN_ZOOM = 1;
+export const MAX_ZOOM = 24;
+export const MAX_AXIS_PRECISION_ZOOM = 6;
+export const TIMING_SAMPLES_PER_BEAT = 48;
+export const MIN_TIMELINE_WIDTH = 1400;
+export const MAX_TIMELINE_WIDTH = 100000;
+export const MAX_TIMELINE_CANVAS_TILE_WIDTH = 4096;
+export const CIRCLE_OBJECT_RADIUS = 16.0;
+export const TAIKO_CIRCLE_RADIUS = 12.0;
+export const TAIKO_FINISHER_CIRCLE_RADIUS = 16.0;
+export const TAIKO_SPINNER_RADIUS = 10.0;
+export const REVERSE_ARROW_ICON_SIZE = 7.0;
+export const TAIKO_DRUMROLL_COLOR = 'rgb(252,191,31)';
+export const TAIKO_SPINNER_COLOR = 'rgb(125,135,150)';
+export const TIMELINE_INTERVAL_STEPS_MS = [
+  1000, 2000, 3000, 5000, 10000, 15000, 30000, 60000, 120000, 300000,
+];
