+++ electron-app/src/components/overview/objects/types.ts
@@ -1,0 +1,6 @@
+import type { Mode, ObjectsOverviewDifficulty } from '../../../Types';
+
+export type ObjectsModeGroup = {
+  mode: Mode;
+  difficulties: ObjectsOverviewDifficulty[];
+};
