+++ electron-app/src/utils/color.ts
@@ -1,0 +1,37 @@
+export function withAlpha(color: string, alpha: number) {
+  const safeAlpha = Math.max(0, Math.min(1, alpha));
+
+  if (color.startsWith('#')) {
+    const normalized =
+      color.length === 4
+        ? `#${color[1]}${color[1]}${color[2]}${color[2]}${color[3]}${color[3]}`
+        : color;
+    const red = Number.parseInt(normalized.slice(1, 3), 16);
+    const green = Number.parseInt(normalized.slice(3, 5), 16);
+    const blue = Number.parseInt(normalized.slice(5, 7), 16);
+    return `rgba(${red}, ${green}, ${blue}, ${safeAlpha})`;
+  }
+
+  if (color.startsWith('rgba(')) {
+    return color.replace(/rgba\(([^)]+),\s*[^,]+\)$/, `rgba($1, ${safeAlpha})`);
+  }
+
+  if (color.startsWith('rgb(')) {
+    return color.replace('rgb(', 'rgba(').replace(')', `, ${safeAlpha})`);
+  }
+
+  return color;
+}
+
+export function hexToRgb(hex: string) {
+  const normalized = hex.replace('#', '');
+  return {
+    r: Number.parseInt(normalized.slice(0, 2), 16),
+    g: Number.parseInt(normalized.slice(2, 4), 16),
+    b: Number.parseInt(normalized.slice(4, 6), 16),
+  };
+}
+
+export function clampColor(value: number) {
+  return Math.max(0, Math.min(255, Math.round(value)));
+}
