+++ electron-app/src/utils/formatters.ts
@@ -1,0 +1,7 @@
+export function formatNullable(value: string | number | null | undefined, fallback = 'N/A') {
+  if (value === null || value === undefined || value === '') {
+    return fallback;
+  }
+
+  return String(value);
+}
